+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Next release
 * Add support for `columns` and `additional_columns` on Report creation
+* Add `declaraction` attribute in CustomsInfo class
 
 ## v5.2.0 (2022-02-25)